Add IProjectile.{Damage, Knockback}

+++ src/Orion.Core/Items/IItem.cs
@@ -45,19 +45,19 @@ namespace Orion.Core.Items
         /// Gets or sets the item's stack size.
         /// </summary>
         /// <value>The item's stack size.</value>
-        public short StackSize { get; set; }
+        public int StackSize { get; set; }
 
         /// <summary>
         /// Gets or sets the item's damage.
         /// </summary>
         /// <value>The item's damage.</value>
-        public short Damage { get; set; }
+        public int Damage { get; set; }
 
         /// <summary>
         /// Gets or sets the item's use time: i.e., the number of ticks before which the item can be used again.
         /// </summary>
         /// <value>The item's use time.</value>
-        public short UseTime { get; set; }
+        public int UseTime { get; set; }
 
         /// <summary>
         /// Sets the item's <paramref name="id"/>. This will update the item accordingly. 
@@ -91,7 +91,7 @@ namespace Orion.Core.Items
                 throw new ArgumentNullException(nameof(item));
             }
 
-            return new ItemStack(item.Id, item.Prefix, item.StackSize);
+            return new ItemStack(item.Id, item.Prefix, (short)item.StackSize);
         }
     }
 }
